Refactor template import statements and add UserBoxPage.get_for_user method
