+++ backend/botany/apps.py
@@ -4,3 +4,4 @@ from django.apps import AppConfig
 class BotanyConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'botany'
+    label = 'botany'
