+++ backend/botany/models.py
@@ -1,76 +1,13 @@
 import uuid
 from django.db import models
-from django.shortcuts import get_object_or_404
 from django.utils.text import slugify
 from django.utils.translation import gettext_lazy as _
 from modelcluster.models import ClusterableModel
 from modelcluster.fields import ParentalKey
-from wagtail.models import Page, TranslatableMixin, PreviewableMixin, Orderable
-from wagtail.fields import RichTextField
+from wagtail.models import TranslatableMixin, PreviewableMixin, Orderable
 from wagtail.search import index
-from wagtail.admin.panels import FieldPanel, TabbedInterface, TitleFieldPanel, ObjectList
-from wagtail.contrib.routable_page.models import RoutablePageMixin, path
 
 from base.models import GalleryImageMixin
-
-
-class UserBoxPage(RoutablePageMixin, Page):
-    """
-    Represents a box page that can contain multiple plants.
-    """
-    description = RichTextField(
-        blank=True
-    )
-    uuid = models.UUIDField(
-        default=uuid.uuid4,
-        editable=False,
-        unique=True,
-        db_index=True
-    )
-
-    parent_page_types = [
-        'home.UserIndexPage'
-    ]
-    child_page_types = []
-
-    content_panels = [
-        TitleFieldPanel('title', classname="title"),
-        FieldPanel('slug'),
-        FieldPanel('description'),
-    ]
-
-    edit_handler = TabbedInterface([
-        ObjectList(content_panels, heading='Details'),
-        ObjectList(Page.promote_panels, heading='Promote', permission="superuser"),
-        ObjectList(Page.settings_panels, heading='Settings', permission="superuser"),
-    ])
-
-    @path('<slug:plant_slug>/')
-    def plant_details(self, request, plant_slug):
-        plant = get_object_or_404(Plant, slug=plant_slug, box=self)
-        return self.render(
-            request,
-            context_overrides={'plant': plant},
-            template="botany/user_plant.html"
-        )
-
-    def get_context(self, request, *args, **kwargs):
-        context = super().get_context(request, *args, **kwargs)
-        context['plants'] = self.get_plants()
-        return context
-
-    def get_plants(self):
-        return Plant.objects.filter(box=self)
-
-    def get_parent_collection(self):
-        return self.owner.index_collection.collection
-
-    def __str__(self):
-        return self.title
-
-    class Meta:
-        verbose_name = _('box')
-        verbose_name_plural = _('boxes')
 
 
 class Plant(
@@ -81,7 +18,7 @@ class Plant(
     PreviewableMixin
 ):
     box = models.ForeignKey(
-        UserBoxPage,
+        'inventory.InventoryBoxPage',
         related_name='plants',
         on_delete=models.CASCADE
     )
@@ -91,6 +28,9 @@ class Plant(
     )
     description = models.TextField(
         blank=True
+    )
+    quantity = models.PositiveIntegerField(
+        default=1
     )
     date = models.DateField(
         null=True,
