--- backend/botany/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path, include
-
-from .api import router
-
-urlpatterns = [
-    path('v1/', include(router.urls)),
-]
