+++ backend/home/apps.py
@@ -6,6 +6,3 @@ class HomeConfig(AppConfig):
     name = 'home'
     verbose_name = _("Home")
     default_auto_field = 'django.db.models.BigAutoField'
-
-    def ready(self):
-        import home.signals  # noqa: F401
