+++ backend/inventory/serializers.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+from .models import InventoryBoxPage
+
+
+class InventoryBoxSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = InventoryBoxPage
+        fields = ['id', 'name', 'description', 'collection']
