+++ backend/inventory/templates/inventory/base.html
@@ -1,0 +1,8 @@
+{% extends "base/base.html" %}
+{% load static %}
+
+{% block extra_css %}
+  <link rel="stylesheet" type="text/css" href="{% static 'inventory/inventory.css' %}">
+{% endblock  %}
+
+{% block content %}{% endblock %}
