+++ backend/inventory/templates/inventory/includes/featured.html
