+++ backend/inventory/widgets.py
@@ -1,0 +1,3 @@
+from .viewsets import box_chooser_viewset
+
+BoxChooserWidget = box_chooser_viewset.widget_class
